+++ src/gui/display.py
@@ -44,3 +44,5 @@ class Display():
                         self.surface (pygame.surface.Surface): surface object
         """
         return self.surface
+    
+