+++ src/gui/image.py
@@ -1,4 +1,4 @@
-import pygame
+import pygame as pg
 
 
 class Image():
@@ -26,32 +26,38 @@ class Image():
                             ON (bool): by default True, turns the gameobject on
         """
         self.screen = screen
+        self.file = file
+        self.col, self.row = col, row
+
         self.scale_x, self.scale_y = scale_x, scale_y
-        self.image = pygame.image.load(file)
-        self.image = pygame.transform.scale(self.image,
+        self.image = pg.image.load(file)
+        self.image = pg.transform.smoothscale(self.image,
                                             (self.image.get_width()*scale_x,
                                             self.image.get_height()*scale_y),)
+        self.scale_x, self.scale_y = self.image.get_width()/self.screen.get_width(),self.image.get_height()/self.screen.get_height()
+
         self.rect = self.image.get_rect()
         self.rect.center = (screen.get_width()*col, screen.get_height()*row)
 
-        # This special variable is responsible of actually drawing the image
-        # if it is not True then the image is not drawn and gameobject functionality
-        # will cease.
-
-        # It is the keystone on which scene management is built in my version of pygame
-        # scenes are basically just lists of gameobjects that are either on or off.
-        # all objects are initially on
-        self.is_on = True
-
-    def draw(self, col, row):
+    def draw(self):
         """
-        Draws the Image in game.
+        Draws the Image in game based on the values of self.rect
+        """
+        self.rect.center = (self.screen.get_width()*self.col, self.screen.get_height()*self.row)
+        self.screen.blit(self.image, (self.rect.x, self.rect.y))
+
+    def resize(self, width, height):
+        """
+        Called upon VIDEORESIZE event, where loads a new image from assets and
+        resizes it according to screen parameters. This is done to keep the image
+        quality good.
 
                 Parameters:
-                    screen (pygame.surface.Surface): the screen object used in drawing
-                    col (int): defines x coordinate through operation self.screen.getwidth()*x
-                    row (int): defines y coordinate through operation self.screen.getheight()*y
+                        width: screen width
+                        height: screen height
         """
-        col, row = self.screen.get_width()*col, self.screen.get_height()*row
-        self.rect.center = (col, row)
-        self.screen.blit(self.image, (self.rect.x, self.rect.y))
+        w, h = width*self.scale_x, height*self.scale_y
+        self.image = pg.transform.smoothscale(pg.image.load(self.file), 
+                                        (w,h),)
+        self.rect.size = w,h
+        self.image.get_rect().center = (self.screen.get_width()*self.col, self.screen.get_height()*self.row)