+++ documentation/changelog.md
@@ -7,3 +7,10 @@
 - Display made at gui/display
 - Downloaded some menu assets to assets/menu_items/menu
 - Some other miscallaneus additions + documentation
+
+## Week 4
+
+- Created the basic menu structure with all requisite transitions par from game -> settings and game -> menu
+- Fleshed out the button class, and made it inherit off of the Image class
+- Cleaned up the functionality of GameOfLife by a lot
+- Downloaded some assets that will probably be used in the actual implemenation of the game